requirements.txt
# internalstate_code dependencies
# Optional: use a venv (e.g. iblenv) and pip install -r requirements.txt

# Data & IBL
numpy==1.23.5
pandas==2.2.2
ONE-api==2.8.1
pyarrow==17.0.0

# Preprocessing / VAE dataset (scipy, sklearn)
scipy==1.14.0
scikit-learn==1.4.2

# VAE training and latents (TensorFlow includes Keras)
tensorflow>=2.12,<2.16

# Figures / analysis (if running manuscript scripts)
matplotlib==3.9.1
seaborn==0.13.2
psychofit  # for fig6 psychometric curves
